requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.3.3
newspaper3k==0.2.8
nltk==3.4
Pillow==5.4.1
pycountry==18.12.8
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
requests-file==1.4.3
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.24.1
